refactor(throttle): split TokenBucket runtime shell from pure plan_consume kernel
